Function to determine the depth to bedrock based on litholog intervals that are labelled as either 'Bedrock' or 'Surficial'

This function finds the first occurrence of 'Bedrock' that is below any other intervals labelled as 'Surficial' in each log. As such, this represents a maximum estimate of bedrock depth; it avoids the problems with glaciotectonic rafts of bedrock, but can potentially overestimate bedrock depth if some intervals are misclassified as surficial with the bedrock strata.

Usage

pick_bedrock(lithologs, response = ".pred_class", option = c("last", "first"))

Arguments

lithologs

data.table of litholog data containing the column '.pred_class' which as two factor levels, 'Bedrock' and 'Surficial'.

response

character, name of the response variable column, default is '.pred_class'.

option

character, one of c("last", "first"). Indicates whether to define the top of the bedrock based on the lowermost occurrence of predicted bedrock that occurs under all other predicted surficial units, or the uppermost occurrence of predicted bedrock. The default is "last". This makes the predicted bedrock top less susceptible to the presence of rafted bedrock rafts, although in some cases could make the predicted bedrock top too deep, if units in the bedrock have been misclassified as surficial.

Value

data.table containing the bedrock depths per well, with 'gicwellid' and '.bedrock_dep' columns.